Fix: Render player models correctly on parking sectors
The player models were not being rendered correctly on parking sectors. This commit fixes this
issue by:

- Creating a new function `calculatePlayerPositionOnParking` to calculate the position of
player models on parking sectors. - Creating a new function `getPlayerRotationOnSector` to
calculate the rotation of player models on parking sectors. - Using the new functions in the
`GameBoard` component to render player models correctly on parking sectors.