Add receive scores for active cards instant card
This card gives the player points equal to twice the number of their active cards. If the
player has no active cards, they get to reroll the wheel.

Also fix a bug where the building bonus card description would not show on mobile.

Also fix a bug where the games history page would crash if the history data was not loaded yet.